# [Enhancement] optimize binary to decimal converter in parquet