@dirtysalt 2022-10-09
run starrocks_admit_test